feat: implement session logic module
